2024.04.15

+++ src/main/java/org/example/qdunavigation/pojo/Node.java
@@ -2,8 +2,8 @@ package org.example.qdunavigation.pojo;
 
 public class Node {
     private String node_no;
-    private int pos_x;
-    private int pos_y;
+    private double pos_x;
+    private double pos_y;
     public String getNode_no() {
         return node_no;
     }
@@ -12,19 +12,19 @@ public class Node {
         this.node_no = node_no;
     }
 
-    public int getPos_x() {
+    public double getPos_x() {
         return pos_x;
     }
 
-    public void setPos_x(int pos_x) {
+    public void setPos_x(double pos_x) {
         this.pos_x = pos_x;
     }
 
-    public int getPos_y() {
+    public double getPos_y() {
         return pos_y;
     }
 
-    public void setPos_y(int pos_y) {
+    public void setPos_y(double pos_y) {
         this.pos_y = pos_y;
     }
     @Override
